requirements.txt
openai
langchain
streamlit
chromadb
python-dotenv
tiktoken
